+++ internal/runner/runner.go
@@ -1,18 +1,13 @@
 package runner
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
-	"net/http"
 	"os"
 	"os/signal"
-	"regexp"
 
-	"github.com/acarl005/stripansi"
 	"github.com/logrusorgru/aurora"
 	"github.com/projectdiscovery/gologger"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/remeh/sizedwaitgroup"
 	"github.com/satyrius/gonx"
 	"ktbs.dev/teler/common"
@@ -22,31 +17,14 @@ import (
 	"ktbs.dev/teler/pkg/teler"
 )
 
-func removeLBR(s string) string {
-	re := regexp.MustCompile(`\x{000D}\x{000A}|[\x{000A}\x{000B}\x{000C}\x{000D}\x{0085}\x{2028}\x{2029}]`)
-	return re.ReplaceAllString(s, ``)
-}
-
 // New read & pass stdin log
 func New(options *common.Options) {
-	var input *os.File
-	var out string
-	var pass int
+	var (
+		input *os.File
+		pass  int
+	)
 
-	metric, promserve, promendpoint := prometheus(options)
-	if metric {
-		go func() {
-			http.Handle(promendpoint, promhttp.Handler())
-
-			err := http.ListenAndServe(promserve, nil)
-			if err != nil {
-				errors.Exit(err.Error())
-			}
-		}()
-
-		metrics.Init()
-		gologger.Info().Msgf("Listening metrics on http://" + promserve + promendpoint)
-	}
+	go metric(options)
 
 	jobs := make(chan *gonx.Entry)
 	gologger.Info().Msg("Analyzing...")
@@ -64,48 +42,25 @@ func New(options *common.Options) {
 	con := options.Concurrency
 	swg := sizedwaitgroup.New(con)
 	go func() {
-		for log := range jobs {
+		for job := range jobs {
 			swg.Add()
 			go func(line *gonx.Entry) {
 				defer swg.Done()
 
 				threat, obj := teler.Analyze(options, line)
-
 				if threat {
-					if metric {
-						metrics.GetThreatTotal.WithLabelValues(obj["category"]).Inc()
-					}
-
-					if options.JSON {
-						json, err := json.Marshal(obj)
-						if err != nil {
-							errors.Exit(err.Error())
-						}
-						out = fmt.Sprintf("%s\n", json)
-					} else {
-						out = fmt.Sprintf("[%s] [%s] [%s] %s\n",
-							aurora.Cyan(obj["time_local"]),
-							aurora.Green(obj["remote_addr"]),
-							aurora.Yellow(obj["category"]),
-							aurora.Red(obj[obj["element"]]),
-						)
-					}
-
-					fmt.Print(out)
-
-					if options.Output != "" {
-						if !options.JSON {
-							out = stripansi.Strip(out)
-						}
-
-						if _, write := options.OutFile.WriteString(out); write != nil {
-							errors.Show(write.Error())
-						}
-					}
+					fmt.Printf("[%s] [%s] [%s] %s\n",
+						aurora.Cyan(obj["time_local"]),
+						aurora.Green(obj["remote_addr"]),
+						aurora.Yellow(obj["category"]),
+						aurora.Red(obj[obj["element"]]),
+					)
 
 					alert.New(options, common.Version, obj)
+					log(options, obj)
+					metrics.PrometheusInsert(obj)
 				}
-			}(log)
+			}(job)
 		}
 	}()