+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project should be documented in this file.
 
+### v2.0.0-beta
+
+- Add supporting RAW requests from CVEs templates
+- Add Zinc logs engine
+- Refactoring configuration structures
+- Remove `-o/--output` & `--json` flags
+- Add custom threat rules
+
 ### v1.2.2
 
 - Add utility for get datasets